requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
python-dotenv
openai

# Web Framework
fastapi
uvicorn[standard]
python-multipart

# PDF Processing
pypdf2
pdfplumber
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-ollama

